refactor(frontend): split settings normalization flows
