+++ packages/frontend/src/components/features/settings/normalization/global-git.ts
@@ -1,0 +1,5 @@
+import type { GlobalGitConfig } from "@openducktor/contracts";
+
+export const normalizeGlobalGitConfigForSave = (git: GlobalGitConfig): GlobalGitConfig => ({
+  defaultMergeMethod: git.defaultMergeMethod,
+});
